rename aux_server.py to server.py
